requirements.txt
Pillow
streamlit
easyocr
gtts
google-generativeai
